+++ src/main/java/at/hannibal2/skyhanni/data/ApiDataLoader.kt
@@ -52,7 +52,7 @@ class ApiDataLoader {
     }
 
     @SubscribeEvent
-    fun onStatusBar(event: ProfileJoinEvent) {
+    fun onProfileJoin(event: ProfileJoinEvent) {
         currentProfileName = event.name
         updateApiData()
     }